ENHANCEMENT: Organized the files for better EFFICIENCY  - Created views folder for components that take up full pages
