--- src/components/HomePage.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function HomePage() {
-    return (
-        <div>
-            This is the HomePage Component
-        </div>
-    )
-}
