+++ src/views/SigninPage.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function SigninPage() {
+  return <div>ME SIGNIN</div>;
+}
